AMSD-18: move classes from model to model.entities package
