+++ AgriculturalManagement/app/src/main/java/com/example/agriculturalmanagement/model/entities/AggrData.java
@@ -1,4 +1,4 @@
-package com.example.agriculturalmanagement.model;
+package com.example.agriculturalmanagement.model.entities;
 
 import java.time.Duration;
 import java.util.List;
