+++ AgriculturalManagement/app/src/main/java/com/example/agriculturalmanagement/model/entities/LinArea.java
@@ -1,8 +1,7 @@
-package com.example.agriculturalmanagement.model;
+package com.example.agriculturalmanagement.model.entities;
 
 import java.lang.String;
 import java.util.Dictionary;
-import java.util.List;
 
 public class LinArea implements GenArea{
