+++ AgriculturalManagement/app/src/main/java/com/example/agriculturalmanagement/model/entities/POI.java
@@ -1,4 +1,4 @@
-package com.example.agriculturalmanagement.model;
+package com.example.agriculturalmanagement.model.entities;
 
 import java.lang.String;
 import java.lang.Exception;
