+++ AgriculturalManagement/app/src/main/java/com/example/agriculturalmanagement/model/entities/PhysicalAddress.java
@@ -1,4 +1,4 @@
-package com.example.agriculturalmanagement.model;
+package com.example.agriculturalmanagement.model.entities;
 
 import java.lang.String;
 import java.lang.Exception;
